Increase code coverage
